add one class mode

+++ sorce_code/Data_analytics.py
@@ -13,8 +13,8 @@ def get_ttests(df_a, df_b, tgt_clm, name_a='df_a', name_b='df_b', max_ylim=None,
     #Process by column 
     for clm in tgt_clm:
         #Delete NaN
-        data_a = np.array(df_a[clm][~df_a[clm].isnull()])
-        data_b = np.array(df_b[clm][~df_b[clm].isnull()])
+        data_a = np.array(df_a[clm][~df_a[clm].isnull()]).astype('float')
+        data_b = np.array(df_b[clm][~df_b[clm].isnull()]).astype('float')
         
         #Stack ttest result by column
         result_T = stats.ttest_ind(data_a, data_b, equal_var = False)
